minor: Structure and interface refactoring (#30)
* [PAGOPA-2696] fix: updating gitignore
* [PAGOPA-2696] chore: refactoring controller structure
* [PAGOPA-2696] chore: refactoring organization-related controller
* [PAGOPA-2696] fix: updating pre-commit hook
* [PAGOPA-2696] chore: refactoring PSP-related controller
* [PAGOPA-2696] chore: refactoring utility and exceptions packages
* [PAGOPA-2696] chore: refactoring responses from controller
* [PAGOPA-2696] chore: refactoring clients package
