+++ .gitignore
@@ -1,31 +1,51 @@
-#apps
-.idea
+#Maven
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+release.properties
+.flattened-pom.xml
 
-# Compiled class file
-*.class
+# Eclipse
+.project
+.classpath
+.settings/
+bin/
 
-# Log file
-*.log
+# IntelliJ
+.idea/
+*.ipr
+*.iml
+*.iws
+.run/
 
-# BlueJ files
-*.ctxt
+# NetBeans
+nb-configuration.xml
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+# Visual Studio Code
+.vscode
+.factorypath
 
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-replay_pid*
-/target/
-/infra/.terraform/
-/.identity/.terraform/
+# OSX
 .DS_Store
+
+# Vim
+*.swp
+*.swo
+
+# patch
+*.orig
+*.rej
+
+# Local environment
+.env
+
+.cache
+.cache_ggshield
+/.quarkus/
+.identity/.terraform/*
+
+# Helm
+/helm/charts/*
+**/.terraform/
+/.run/
