+++ .pre-commit-config.yaml
@@ -10,11 +10,8 @@ repos:
         language_version: python3
         stages: [ commit ]
 
-  - repo: local
+
+  - repo: https://github.com/pagopa/google-style-precommit-hook
+    rev: be92ca3b151dda05fb027ee738509ebf68786bc1
     hooks:
       - id: google-style-java
-        name: Google Java Code Style for Java
-        description: Formats code in Google's Java codestyle.
-        entry: ./format-code.sh
-        language: script
-        files: \.java$ # We don't technically need this, as the script will filter for us, but this will allow the hook to be skipped if no Java is changed.
