+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/interfaces/IPSPController.java
@@ -1,0 +1,57 @@
+package it.gov.pagopa.fdrtechsupport.controller.interfaces;
+
+import it.gov.pagopa.fdrtechsupport.controller.model.response.FrResponse;
+import jakarta.validation.constraints.NotNull;
+import jakarta.ws.rs.*;
+import org.eclipse.microprofile.openapi.annotations.Operation;
+import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
+import org.eclipse.microprofile.openapi.annotations.tags.Tag;
+
+import java.time.LocalDate;
+
+@Path("/psps")
+@Consumes("application/json")
+@Produces("application/json")
+@Tag(name = "PSP", description = "APIs for inspecting PSP usage of flows")
+public interface IPSPController {
+
+    @GET
+    @Operation(
+            operationId = "PSPController_searchFlowByPsp",
+            summary = "...",
+            description = "...")
+    @APIResponses(value = {})
+    @Path("/{pspId}")
+    FrResponse searchFlowByPsp(
+            @PathParam("pspId") @NotNull String pspId,
+            @QueryParam("dateFrom") LocalDate dateFrom,
+            @QueryParam("dateTo") LocalDate dateTo,
+            @QueryParam("flowId") String flowId,
+            @QueryParam("organizationId") String organizationId);
+
+    @GET
+    @Operation(
+            operationId = "PSPController_searchFlowByPspAndIuv",
+            summary = "...",
+            description = "...")
+    @APIResponses(value = {})
+    @Path("/{pspId}/iuv/{iuv}")
+    FrResponse searchFlowByPspAndIuv(
+            @PathParam("pspId") @NotNull String pspId,
+            @PathParam("iuv") @NotNull String iuv,
+            @QueryParam("dateFrom") LocalDate dateFrom,
+            @QueryParam("dateTo") LocalDate dateTo);
+
+    @GET
+    @Operation(
+            operationId = "PSPController_searchFlowByPspAndIuv",
+            summary = "...",
+            description = "...")
+    @APIResponses(value = {})
+    @Path("/{pspId}/iur/{iur}")
+    FrResponse searchFlowByPspAndIur(
+            @PathParam("pspId") @NotNull String pspId,
+            @PathParam("iur") @NotNull String iur,
+            @QueryParam("dateFrom") LocalDate dateFrom,
+            @QueryParam("dateTo") LocalDate dateTo);
+}
