+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/model/response/FdrFullInfoResponse.java
@@ -1,0 +1,15 @@
+package it.gov.pagopa.fdrtechsupport.controller.model.response;
+
+import java.time.LocalDate;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class FdrFullInfoResponse {
+  private LocalDate dateFrom;
+  private LocalDate dateTo;
+  private String data;
+}
