--- src/main/java/it/gov/pagopa/fdrtechsupport/resources/response/FdrFullInfoResponse.java
@@ -1,15 +1,0 @@
-package it.gov.pagopa.fdrtechsupport.resources.response;
-
-import java.time.LocalDate;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class FdrFullInfoResponse {
-  private LocalDate dateFrom;
-  private LocalDate dateTo;
-  private String data;
-}
