--- src/main/java/it/gov/pagopa/fdrtechsupport/util/AppConstant.java
@@ -1,16 +1,0 @@
-package it.gov.pagopa.fdrtechsupport.util;
-
-public class AppConstant {
-
-  public static final String SERVICE_CODE_APP = "FDR";
-
-  public static final String PSP_ID = "pspId";
-  public static final String EC_ID = "ecId";
-  public static final String FLOW_NAME = "flowName";
-  public static final String INDEXES = "indexes";
-  public static final String INTERNAL_READ = "internalRead";
-
-  private AppConstant() {
-    throw new IllegalStateException("Constants class");
-  }
-}
