--- src/test/java/it/gov/pagopa/fdrtechsupport/resources/GenericTest.java
@@ -1,16 +1,0 @@
-package it.gov.pagopa.fdrtechsupport.resources;
-
-import static io.restassured.RestAssured.given;
-
-import io.quarkus.test.junit.QuarkusTest;
-import jakarta.ws.rs.core.MediaType;
-import org.junit.jupiter.api.Test;
-
-@QuarkusTest
-class GenericTest {
-
-  @Test
-  public void info() {
-    given().when().get("/info").then().statusCode(200).contentType(MediaType.APPLICATION_JSON);
-  }
-}
